Reorganize folder structure for Packages classes
